% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_distance.R
\name{group_by_distance}
\alias{group_by_distance}
\title{Agrupar geometrías por proximidad espacial}
\usage{
group_by_distance(sf, distance)
}
\arguments{
\item{sf}{Objeto de clase \code{sf} o \code{sfc}.}

\item{distance}{Valor numérico. Distancia máxima en metros para considerar dos elementos como conectados.}
}
\value{
Un vector de enteros indicando el ID del grupo (membership) para cada geometría.
}
\description{
Esta función agrupa las geometrías de un objeto \code{sf} que se encuentran dentro de una
distancia umbral especificada. Utiliza componentes conectados en un grafo de adyacencia.
}
\examples{
\dontrun{
  # Agrupar geometrías que distan menos de 50 metros entre sí
  sf_obj$group <- group_by_distance(sf_obj, distance = 50)
  # Agrupar geometrias en base a una agrupación de atributos
  variables <- c("Grupo_1", "Grupo_2")
  sf_obj \%>\% 
   dplyr::group_by(!!dplyr::syms(variables)) \%>\% 
   dplyr::mutate(group = group_by_distance(geometry, distance = 50))
}
}
