% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_ziprar.R
\name{install.7z}
\alias{install.7z}
\title{Instalar 7-Zip}
\usage{
install.7z()
}
\value{
Invisible. El resultado de la función \code{installr::install.URL}.
}
\description{
Descarga e instala automáticamente la versión de 7-Zip correspondiente a la arquitectura
del sistema (32 o 64 bits). Esta herramienta es necesaria para descomprimir archivos \code{.rar}
con las funciones \code{read_sf_ziprar} y \code{copy_sf_ziprar}.
}
\examples{
\dontrun{
  install.7z()
}
}
