% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio_elqui.R
\docType{data}
\name{rio_elqui}
\alias{rio_elqui}
\title{Rio Elqui}
\format{
\subsection{\code{rio_elqui}}{

Simple feature con 1 fila y 3 columnas:
\describe{
\item{nombre}{Nombre del curso de agua}
\item{tipo}{tipo de curso de agua}
\item{strahler_n}{Orden de corriente Strahler}
}
}
}
\source{
\url{https://geoportal.cl/geoportal/catalog/36436/Hidrograf\%C3\%ADa\%20de\%20la\%20regi\%C3\%B3n\%20de\%20Arica\%20a\%20la\%20regi\%C3\%B3n\%20de\%20Los\%20Lagos}
}
\usage{
rio_elqui
}
\description{
Objeto sf del Rio Elqui, Chile.
}
\keyword{datasets}
