% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_add_geometry.R
\name{st_add_coords}
\alias{st_add_coords}
\title{Agregar coordenadas X e Y a un objeto sf}
\usage{
st_add_coords(sf, names = c("UTM_E", "UTM_N"))
}
\arguments{
\item{sf}{Objeto de clase \code{sf}. Debe contener geometrías de tipo POINT.}

\item{names}{Vector de caracteres de longitud 2. Nombres para las columnas de coordenadas (X, Y).
Por defecto c("UTM_E", "UTM_N").}
}
\value{
El objeto \code{sf} original con dos columnas nuevas correspondientes a las coordenadas.
}
\description{
Extrae las coordenadas de la geometría (asumiendo geometrías de tipo POINT) y las agrega
como nuevas columnas al objeto sf.
}
\examples{
\dontrun{
 # Crea los campos 'UTM_E' y 'UTM_N' por defecto
 sf_puntos \%>\% st_add_coords(sf_puntos)
 # Asignar nombre a los campos con las coordenadas
 sf_puntos \%>\% st_add_coords(sf_puntos, names = c("Coord_X", "Coord_Y"))
}
}
