% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_add_geometry.R
\name{st_add_sup_ha}
\alias{st_add_sup_ha}
\title{Agregar superficie en hectáreas}
\usage{
st_add_sup_ha(sf, digits = 2)
}
\arguments{
\item{sf}{Objeto de clase \code{sf}.}

\item{digits}{Entero. Número de decimales para redondear. Por defecto 2.}
}
\value{
El objeto \code{sf} con la columna 'Sup_ha'.
}
\description{
Calcula el área de las geometrías y agrega una columna 'Sup_ha' con el valor en hectáreas.
}
\examples{
\dontrun{
 sf_poligonos <- st_add_sup_ha(sf_poligonos)
 sf_poligonos <- st_add_sup_ha(sf_poligonos, digit = 0)
}
}
