% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_add_geometry.R
\name{st_add_sup_m2}
\alias{st_add_sup_m2}
\title{Agregar superficie en metros cuadrados}
\usage{
st_add_sup_m2(sf, digits = 0)
}
\arguments{
\item{sf}{Objeto de clase \code{sf}.}

\item{digits}{Entero. Número de decimales para redondear. Por defecto 0.}
}
\value{
El objeto \code{sf} con la columna 'Sup_m2'.
}
\description{
Calcula el área de las geometrías y agrega una columna 'Sup_m2' con el valor en metros cuadrados.
}
\examples{
\dontrun{
 sf_poligonos <- st_add_sup_m2(sf_poligonos)
 sf_poligonos <- st_add_sup_m2(sf_poligonos, digits = 1)
}
}
