% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_is_utm.R
\name{st_is_utm}
\alias{st_is_utm}
\title{Verificar si un objeto sf tiene proyección UTM}
\usage{
st_is_utm(sf)
}
\arguments{
\item{sf}{Objeto de clase 'sf' o 'sfc'}
}
\value{
Logical. TRUE si el CRS es UTM, FALSE en caso contrario.
}
\description{
Verificar si un objeto sf tiene proyección UTM
}
\examples{
\dontrun{
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
st_is_utm(nc)
}
}
