% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_position.R
\name{st_position}
\alias{st_position}
\title{Posición relativa a una geometría de referencia}
\usage{
st_position(x, y, rel = c("NS", "EO"))
}
\arguments{
\item{x}{Objeto \code{sf} o \code{sfc} (preferiblemente un punto) cuya posición se quiere determinar.}

\item{y}{Objeto \code{sf} o \code{sfc} que sirve como línea o polígono de referencia.}

\item{rel}{Cadena de caracteres. Define el eje de comparación: "NS" para
Norte-Sur (compara coordenadas Y) o "EO" para Este-Oeste (compara coordenadas X).}
}
\value{
Una cadena de caracteres: "Norte", "Sur", "Este", "Oeste" o "Dentro" cuando se calcula la posición respecto de un polígono.
}
\description{
Calcula la posición cardinal (Norte/Sur o Este/Oeste) de un objeto sf ('x')
con respecto al punto más cercano de otro objeto sf de referencia ('y').
}
\examples{
\dontrun{
  library(dplyr)
  # Posición de un punto respecto del rio Elqui
  punto <- sf::st_sample(sf::st_buffer(rio_elqui, 100), 1)
  st_position(punto, rio_elqui)
  st_position(sf::st_buffer(punto, 500), rio_elqui)

  # posición de un punto respecto dela región de Valparaíso
  punto_valpo <- sf::st_sample(sf::st_buffer(valparaiso, 0), 1)
  st_position(punto_valpo, valparaiso)

  # posición de un punto respecto dela región de Valparaíso
  st_position(sf::st_buffer(punto_valpo, 20000), valparaiso)
  st_position(sf::st_buffer(punto_valpo, 20000), rio_elqui)
}
}
