% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_redim.R
\name{st_redim_stars}
\alias{st_redim_stars}
\title{Redimensionar objeto stars}
\usage{
st_redim_stars(x)
}
\arguments{
\item{x}{Un objeto de clase \code{stars}.}
}
\value{
Un objeto de clase \code{stars} con las dimensiones actualizadas.
}
\description{
Actualiza las dimensiones espaciales ('x' e 'y') de un objeto \code{stars}.
Calcula el nuevo \code{offset} considerando el índice de inicio (\code{from}) y la resolución (\code{delta}),
lo cual es útil para corregir la extensión espacial después de realizar recortes o subsets
que no actualizan automáticamente estos metadatos.
}
\examples{
\dontrun{
  # x <- stars::read_stars("archivo.tif")
  # y <- sf::read_sf("poligono.shp") \%>\% sf::st_transform(sf::st_crs(x))
  # x_y <- x[y]
  # x_redim <- st_redim_stars(x_y)
}
}
