% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_union.R
\name{st_union_features}
\alias{st_union_features}
\title{Unir características de dos capas sf (Intersección + Diferencia)}
\usage{
st_union_features(x, y)
}
\arguments{
\item{x}{Objeto \code{sf} (capa base).}

\item{y}{Objeto \code{sf} (capa de superposición).}
}
\value{
Un objeto \code{sf} que contiene las geometrías resultantes.
}
\description{
Realiza una operación espacial que combina la intersección de \code{x} e \code{y}
con la diferencia de \code{x} menos \code{y}. El resultado cubre espacialmente la misma
área que \code{x}, pero con las geometrías divididas por los límites de \code{y} y
los atributos de \code{y} agregados en las zonas de intersección.
}
\examples{
\dontrun{
  # res <- st_union_features(poligonos_a, poligonos_b)
}
}
