% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valparaiso.R
\docType{data}
\name{valparaiso}
\alias{valparaiso}
\title{valparaiso}
\format{
\subsection{\code{valparaiso}}{

Simple feature con 1 fila y 3 columnas:
\describe{
\item{CUT_REG}{Código Único Territorial}
\item{REGION}{Nombre de la Región}
\item{SUPERFICIE}{Superficie en km2}
}
}
}
\source{
\url{https://geoportal.cl/geoportal/catalog/36391/Divisi\%C3\%B3n\%20Pol\%C3\%ADtica\%20Administrativa\%202023}
}
\usage{
valparaiso
}
\description{
Objeto sf de la Región de Valparaíso, Chile.
}
\keyword{datasets}
